src/test/java/com/social/media/controller/UserControllerTest.java
package com.social.media.controller;

import com.social.media.entity.User;
import com.social.media.service.UserService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class UserControllerTest {

    @Mock
    private UserService userService;

    @InjectMocks
    private UserController userController;


    @Test
    void testFollowUser() {
        User user = new User();
        user.setId(1L);
        user.setUsername("user1");

        when(userService.followUser(1L, 2L)).thenReturn(user);

        ResponseEntity<User> response = userController.followUser(1L, 2L);
        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertNotNull(response.getBody());
        assertEquals(1L, response.getBody().getId());
    }

}